Revert "Updated views and increased max rating"
This reverts commit 7c6352c580b153becf904bc26d7ac0cf6099de3d.

+++ models/review.js
@@ -10,7 +10,7 @@ var Review = db.define('review', {
     type: Sequelize.STRING(10000)
   },
   rating: {
-    type: Sequelize.INTEGER(100)
+    type: Sequelize.INTEGER(5)
   }
 });
